refactor: simplify auto_encoder to autoencoder;
